DIP/requirements.txt
streamlit
Pillow
numpy
opencv-python
rembg
